+++ rust/operator-binary/src/main.rs
@@ -23,7 +23,8 @@ use stackable_operator::{
     },
 };
 use stackable_superset_crd::{
-    commands::{DruidConnection, SupersetDB},
+    druidconnection::DruidConnection,
+    supersetdb::SupersetDB,
     SupersetCluster,
 };
 
@@ -160,9 +161,9 @@ async fn main() -> anyhow::Result<()> {
                             .state()
                             .into_iter()
                             .filter(move |druid_connection| {
-                                &druid_connection.spec.superset_db_namespace
+                                &druid_connection.spec.superset_cluster_namespace
                                     == sdb.metadata.namespace.as_ref().unwrap()
-                                    && &druid_connection.spec.superset_db_name
+                                    && &druid_connection.spec.superset_cluster_name
                                         == sdb.metadata.name.as_ref().unwrap()
                             })
                             .map(|druid_connection| ObjectRef::from_obj(&druid_connection))
